[PATRO23-350] change return type when adding category
